rename bit-json to bit-config
