+++ src/consumer/component-ops/inject-conf.js
@@ -2,7 +2,7 @@
 import path from 'path';
 import R from 'ramda';
 import type ConsumerComponent from '../component/consumer-component';
-import ComponentBitConfig from '../bit-json';
+import ComponentBitConfig from '../bit-config';
 import removeEmptyDir from '../../utils/fs/remove-empty-dir';
 import GeneralError from '../../error/general-error';
 import type BitMap from '../bit-map';
